akka.sample.osgi.dining-hakkers integration test
