+++ akka-samples/akka-sample-osgi-dining-hakkers/README.md
