+++ akka-samples/akka-sample-osgi-dining-hakkers/api/src/main/scala/akka/osgi/sample/api/DiningHakkersMessages.scala
