+++ akka-samples/akka-sample-osgi-dining-hakkers/assembly-dist/src/main/distribution/README
