+++ akka-samples/akka-sample-osgi-dining-hakkers/assembly-dist/src/main/distribution/etc/org.ops4j.pax.url.mvn.cfg.onlylocal
