+++ akka-samples/akka-sample-osgi-dining-hakkers/assembly-dist/src/main/resources/etc/readme.txt
