+++ akka-samples/akka-sample-osgi-dining-hakkers/core/src/main/resources/application.conf
