+++ akka-samples/akka-sample-osgi-dining-hakkers/core/src/main/scala/akka/osgi/sample/internal/Table.scala
@@ -1,0 +1,26 @@
+/*
+Copyright 2013 Crossing-Tech
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+   limitations under the License.
+ */
+package akka.sample.osgi.internal
+
+import akka.actor.{ Props, Actor }
+
+class Table extends Actor {
+  val chopsticks = for (i ← 1 to 5) yield context.actorOf(Props[Chopstick], "Chopstick" + i)
+
+  def receive = {
+    case x: Int ⇒ sender ! (chopsticks(x), chopsticks((x + 1) % 5))
+  }
+}
