--- api/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>parent</artifactId>
-        <groupId>com.typesafe.akka.akka-sample.akka-sample-osgi-dining-hakkers</groupId>
-        <version>2.2.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>api</artifactId>
-    <name>Dining Hakker :: Service API</name>
-    <packaging>bundle</packaging>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>com.typesafe.akka</groupId>
-            <artifactId>akka-osgi_${scala.dep.version}</artifactId>
-        </dependency>
-
-    </dependencies>
-</project>
