--- uncommons/pom.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>parent</artifactId>
-        <groupId>com.typesafe.akka.akka-sample.akka-sample-osgi-dining-hakkers</groupId>
-        <version>2.2.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <!--
-    Note that 1.2.3 of uncommons-maths has OSGi meta-data, see https://www.assembla.com/spaces/akka/tickets/2990
-    Once that is available, this module can be removed.
-    @see ticket #2990
-    -->
-
-    <artifactId>uncommons</artifactId>
-    <name>org.uncommons.maths.random</name>
-    <version>1.2.2</version>
-    <packaging>bundle</packaging>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.uncommons.maths</groupId>
-            <artifactId>uncommons-maths</artifactId>
-            <version>1.2.2</version>
-        </dependency>
-        <dependency>
-            <groupId>jfree</groupId>
-            <artifactId>jfreechart</artifactId>
-            <version>1.0.13</version>
-        </dependency>
-        <dependency>
-            <groupId>jfree</groupId>
-            <artifactId>jcommon</artifactId>
-            <version>1.0.16</version>
-        </dependency>
-
-    </dependencies>
-    <build>
-
-        <plugins>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <configuration>
-                    <instructions>
-                        <Bundle-Name>${project.name}</Bundle-Name>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Export-Package>org.uncommons.maths.random</Export-Package>
-                        <Import-Package>!sun.misc, *</Import-Package>
-                        <Private-Package>org.uncommons.maths.binary, org.uncommons.maths, org.uncommons.maths.number</Private-Package>
-                    </instructions>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
